+++ pallets/reversible-transfers/src/tests/mod.rs
@@ -2,7 +2,7 @@
 mod mock;
 
 #[cfg(test)]
-mod tests;
+mod test_reversible_transfers;
 
 #[cfg(test)]
 mod test_high_security_account;
